misc/HtmlTestSuite/results/test-0837.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>CSS Test: invalidation of class changes when the sheet the style depends on goes away</title>
<link href="mailto:emilio@crisal.io" rel="author" title="Emilio Cobos Ãlvarez"/>
<link href="https://drafts.csswg.org/selectors-4/#invalid" rel="help"/>
<link href="https://bugzilla.mozilla.org/show_bug.cgi?id=1432850" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<style>
  body { background: green; }
</style>
<style id="style">
  .red { background: red; }
</style>
</head><body class="red">
Should have a green background.
<script>
test(() =&gt; {
  document.body.offsetTop;
  assert_equals(getComputedStyle(document.body).backgroundColor, "rgb(255, 0, 0)");
  document.body.className = "";
  style.remove();
  assert_equals(getComputedStyle(document.body).backgroundColor, "rgb(0, 128, 0)");
}, "Style should be recomputed correctly when the stylesheet it depends on goes away");
</script>
</body>
</html>
